Remote storr

Create a storr that keeps rds-serialised objects on a remote location. This is the abstract interface (which does not do anything useful) but which can be used with file operation driver to store files elsewhere. This is not intended for end-user use so there is no storr_remote function. Instead this function is designed to support external packages that implement the details. For a worked example, see the package tests (helper-remote.R). In the current implementation these build off of the driver_rds driver by copying files to some remote location.

driver_remote(ops, ..., path_local = NULL)

Arguments

ops
A file operations object. See tests for now to see what is required to implement one.

...
Arguments to pass through to driver_rds, including compress, mangle_key, mangle_key_pad and hash_algorithm.

path_local
Path to a local cache. This can be left as NULL, in which case a per-session cache will be used. Alternatively, explicitly set to a path and the cache can be reused over sessions. Only storr values (i.e., objects) are cached - the key-to-value mapping is always fetched from the remote storage.