Recycle key and namespace

Utility function for driver authors

Usage

join_key_namespace(key, namespace)

Arguments

key

A vector of keys

namespace

A vector of namespace

Value

A list with elements n, key and namespace

Details

This exists to join, predictably, keys and namespaces for operations like mget. Given a vector or scalar for key and namespace we work out what the required length is and recycle key and namespace to the appropriate length.